requirements.txt
absl-py==0.14.1
aiohttp==3.7.4.post0
alpha-vantage==2.3.1
altgraph==0.17
asgiref==3.4.1
astunparse==1.6.3
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==19.3.0
autopy==4.0.0
backcall==0.1.0
beautifulsoup4==4.10.0
bleach==3.1.0
blosc==1.10.6
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
clang==5.0
click==8.0.0
cmake==3.18.4.post1
color-harmony==1.0.1
colorama==0.4.3
cx-Freeze==6.6
cx-Logging==3.0
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
Django==3.2.6
django-tinymce==3.4.0
dlib==19.22.0
docutils==0.17.1
entrypoints==0.3
et-xmlfile==1.1.0
flask-2-ui5-py==0.0.1
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.41.0
gspread==3.7.0
gunicorn==20.1.0
h5py==2.10.0
httplib2==0.19.1
idna==2.10
imageai==2.1.6
importlib-metadata==4.0.1
imutils==0.5.3
iniconfig==1.1.1
install==1.3.4
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==2.0.1
jedi==0.16.0
Jinja2==3.0.1
joblib==0.14.1
jsonschema==3.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==3.0.0
toml==0.10.2
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wikipedia==1.4.0
wrapt==1.12.1
xlrd==2.0.1
yarl==1.7.0
youtube-dl==2021.6.6
zipp==3.4.1